+++ apps/web/changelog.md
@@ -1,3 +1,18 @@
+### [0.0.2]
+- Added Blog Page
+- Added Blog Detail Page
+- Added Blog Post Data Structure
+- Added Comment Data Structure
+- Added Blog Post Navigation
+- Added Blog Post Comments
+- Added Blog Post Likes
+- Added Blog Post Related Posts
+- Added Blog Post Author
+- Added Blog Post Tags
+- Added Blog Post Social Sharing
+- Added Blog Post Save/Bookmark
+- Added Blog Post Newsletter CTA
+
 ### [0.0.1]
 - Intial App Setup
 - Home Page Added